refactor: use Emotion for search bar styling
