--- src/react/table-app/option-bar/search-bar/styles.css
@@ -1,7 +1,0 @@
-.NLT__search input {
-    width: 175px;
-    border-top: 0 !important;
-    border-left: 0 !important;
-    border-right: 0 !important;
-    border-bottom: 1px solid var(--background-modifier-border) !important;
-}
